Disallow writing to register 0
Stop instructions from actually writing to register 0. It might look
unnecessary at first because we can maybe check at decode if we're
actually going to try and raise and exception or something (if we wanted
to, maybe we just ignore the instruction).

Even if we raised an exception doing this is important because this
allows us to introduce bubbles in the pipeline that operate with
register0.