Fix store buffer
Check whether we're full or not with a variable that counts the number
of entries. We add +1 whenever we assign a new entry and substract 1
when we free one entry.

Fix the loop. We do 'entries' iterations from head to tail instead of
going from head to tail (they can be the same)

Also, comments added in general